# openapi.yaml
openapi: 3.0.0
info:
  title: Ordering API
  version: 1.0.0
paths:
...
  /payments:
    get:
      operationId: payments_get
      summary: Get all payments
      responses:
        '200':
          description: Successfully got payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '500':
          description: Failed to get payments
  '/payments/{order_id}':
    post:
      operationId: payments_create
      summary: Perform a payment
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ObjectId'
      responses:
        '201':
          description: Successfully processed payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Failed to process payment
        '404':
          description: Referenced order does not exist
    delete:
      operationId: payments_delete
      summary: Delete a payment
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ObjectId'
      responses:
        '200':
          description: Successfully deleted payment
        '400':
          description: Failed to delete payment
components:
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
...
    Payment:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        customer_name:
          type: string
        order_name:
          type: string
        order_price:
          type: number